rename: 인증 관련 api 네이밍 변경 및 정리
