+++ src/api/auth/deleteAccount.ts
