+++ src/api/auth/logout.ts
